requirements.txt
alpaca-trade-api>=3.0.2,<4.0.0
pandas>=2.0
numpy>=1.24
requests>=2.28
python-dotenv>=1.0
